split kv.go
