+++ cloudflare/kv/delete.go
@@ -1,0 +1,16 @@
+package kv
+
+import (
+	"github.com/syumai/workers/internal/jsutil"
+)
+
+// Delete deletes key-value pair specified by the key.
+//   - if a network error happens, returns error.
+func (ns *Namespace) Delete(key string) error {
+	p := ns.instance.Call("delete", key)
+	_, err := jsutil.AwaitPromise(p)
+	if err != nil {
+		return err
+	}
+	return nil
+}
